// components/inscription/types.ts
export interface FormData {
  // Academic info
  status: string
  university_id: string
  university_name: string
  formation_type: string
  faculty: string
  field_of_study: string

  // Personal info
  full_name: string
  email: string
  password: string
  phone: string
  date_of_birth: string
  gender: string
}